refactor: extract player-lifecycle cluster from state.py God-Object (#1271)
Pull the PlayerRegistry + PowerUpManager delegation layer out of the
game/state.py God-Object into a new PlayerLifecycleMixin, following the
TTS / leaderboard / media-lights / challenge-delegation extractions.

The cluster is pure pass-through delegation (player dict/leader, lookups
by name/session-id/ws, sessions, reactions, admin, average score,
submitted-aggregate, plus steal targeting/execution and streak/bet
counters). Behavior-preserving: methods moved verbatim, public API of
GameState unchanged via inheritance.

state.py 2347 -> 2269 lines.
