+++ custom_components/beatify/game/state.py
@@ -71,6 +71,7 @@ from .serializers import GameStateSerializer
 from .state_challenge import ChallengeMixin
 from .state_leaderboard import LeaderboardMixin
 from .state_media import MediaControlMixin
+from .state_player import PlayerLifecycleMixin
 from .state_tts import TtsAnnouncerMixin
 
 from .types import RoundAnalytics, _get_decade_label
@@ -141,6 +142,7 @@ class GameState(
     ChallengeMixin,
     LeaderboardMixin,
     MediaControlMixin,
+    PlayerLifecycleMixin,
     TtsAnnouncerMixin,
 ):
     """Manages game state and phase transitions.
@@ -158,6 +160,11 @@ class GameState(
     The challenge-delegation subsystem (Issue #1271 next-increment
     extraction, stacked on the media extraction) lives in
     :class:`~custom_components.beatify.game.state_challenge.ChallengeMixin`.
+
+    The player-lifecycle subsystem (Issue #1271 next-increment extraction:
+    PlayerRegistry + PowerUpManager delegation — player lookups, sessions,
+    reactions, admin, steal/streak/bet pass-throughs) lives in
+    :class:`~custom_components.beatify.game.state_player.PlayerLifecycleMixin`.
     """
 
     def __init__(self, time_fn: Callable[[], float] | None = None) -> None:
@@ -358,24 +365,9 @@ class GameState(
         return self._now()
 
     # ------------------------------------------------------------------
-    # Player registry delegation (keep public interface identical)
+    # Player registry / power-up delegation lives in PlayerLifecycleMixin
+    # (Issue #1271 extraction). See game/state_player.py.
     # ------------------------------------------------------------------
-
-    @property
-    def players(self) -> dict[str, PlayerSession]:
-        """Player dict — delegated to PlayerRegistry."""
-        return self._player_registry.players
-
-    @players.setter
-    def players(self, value: dict[str, PlayerSession]) -> None:
-        self._player_registry.players = value
-
-    @property
-    def leader(self) -> PlayerSession | None:
-        """Get current leader player (cached per state change)."""
-        if not self.players:
-            return None
-        return max(self.players.values(), key=lambda p: p.score)
 
     # ------------------------------------------------------------------
     # RoundManager delegation (keep public interface identical)
@@ -516,26 +508,9 @@ class GameState(
         self._round_manager.metadata_pending = value
 
     # ------------------------------------------------------------------
-    # Power-up delegation properties (keep public interface identical)
+    # Power-up delegation properties live in PlayerLifecycleMixin
+    # (Issue #1271 extraction). See game/state_player.py.
     # ------------------------------------------------------------------
-
-    @property
-    def streak_achievements(self) -> dict[str, int]:
-        """Streak achievement counters."""
-        return self._powerup_manager.streak_achievements
-
-    @streak_achievements.setter
-    def streak_achievements(self, value: dict[str, int]) -> None:
-        self._powerup_manager.streak_achievements = value
-
-    @property
-    def bet_tracking(self) -> dict[str, int]:
-        """Bet outcome counters."""
-        return self._powerup_manager.bet_tracking
-
-    @bet_tracking.setter
-    def bet_tracking(self, value: dict[str, int]) -> None:
-        self._powerup_manager.bet_tracking = value
 
     def get_song_difficulty(self, song_uri: str) -> dict[str, Any] | None:
         """Get song difficulty rating — delegated to StatsService."""
@@ -1030,59 +1005,10 @@ class GameState(
 
         return True
 
-    def get_average_score(self) -> int:
-        """Calculate average score of all current players. Delegates to PlayerRegistry."""
-        return self._player_registry.get_average_score()
-
-    def add_player(
-        self, name: str, ws: web.WebSocketResponse
-    ) -> tuple[bool, str | None]:
-        """Add a player to the game. Delegates to PlayerRegistry."""
-        return self._player_registry.add_player(
-            name, ws, self.phase, self.get_average_score
-        )
-
-    def get_player(self, name: str) -> PlayerSession | None:
-        """Get player by name. Delegates to PlayerRegistry."""
-        return self._player_registry.get_player(name)
-
-    def get_player_by_session_id(self, session_id: str) -> PlayerSession | None:
-        """Get player by session ID. Delegates to PlayerRegistry."""
-        return self._player_registry.get_player_by_session_id(session_id)
-
-    def get_player_by_ws(self, ws: web.WebSocketResponse) -> PlayerSession | None:
-        """Get player by WebSocket connection. Delegates to PlayerRegistry."""
-        return self._player_registry.get_player_by_ws(ws)
-
-    def record_reaction(self, player_name: str, emoji: str) -> bool:
-        """Record a player reaction. Delegates to PlayerRegistry."""
-        return self._player_registry.record_reaction(player_name, emoji)
-
-    def get_steal_targets(self, stealer_name: str) -> list[str]:
-        """Get list of players who can be stolen from (Story 15.3). Delegates to PowerUpManager."""
-        return self._powerup_manager.get_steal_targets(stealer_name, self.players)
-
-    def use_steal(self, stealer_name: str, target_name: str) -> dict[str, Any]:
-        """Execute steal power-up (Story 15.3). Delegates to PowerUpManager."""
-        return self._powerup_manager.use_steal(
-            stealer_name, target_name, self.players, self.phase, self._now()
-        )
-
-    def remove_player(self, name: str) -> None:
-        """Remove player from game. Delegates to PlayerRegistry."""
-        self._player_registry.remove_player(name)
-
-    def clear_all_sessions(self) -> None:
-        """Clear all session mappings for game reset. Delegates to PlayerRegistry."""
-        self._player_registry.clear_all_sessions()
-
-    def get_players_state(self) -> list[dict[str, Any]]:
-        """Get player list for state broadcast. Delegates to PlayerRegistry."""
-        return self._player_registry.get_players_state()
-
-    def all_submitted(self) -> bool:
-        """Check if all connected players have submitted. Delegates to PlayerRegistry."""
-        return self._player_registry.all_submitted()
+    # ------------------------------------------------------------------
+    # Player lifecycle / lookup + power-up delegation lives in
+    # PlayerLifecycleMixin (Issue #1271 extraction). See game/state_player.py.
+    # ------------------------------------------------------------------
 
     def check_all_guesses_complete(self) -> bool:
         """
@@ -1248,10 +1174,6 @@ class GameState(
             "is_first_game": comparison["is_first_game"],
             "message": message_data,
         }
-
-    def set_admin(self, name: str) -> bool:
-        """Mark a player as admin. Delegates to PlayerRegistry."""
-        return self._player_registry.set_admin(name)
 
     def _detect_storefront(self) -> str | None:
         """Determine the user's Apple Music storefront for URI resolution.
